Refactor extract address know-how on common mapper
